+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/data/view/column/ColumnI.java
@@ -50,7 +50,7 @@ public class ColumnI<T> implements Column<T> {
 
 	private Lookup<T> ensureInitializedLookup() {
 		if (lookup.isEmpty()) {
-			lookup = Optional.of(Lookups.lookup(view, attribute));
+			lookup = Optional.of(Lookups.persistedLookup(view, attribute));
 		}
 		return lookup.get();
 	}